- resolve #216: Improve Rollbar.NET integration within the Xamarin sample
