+++ Sample.Xamarin.Forms/Sample.Xamarin.Forms.UWP/App.xaml.cs
@@ -31,12 +31,15 @@ namespace Sample.Xamarin.Forms.UWP
         /// </summary>
         public App()
         {
-            ConfigureRollbarSingleton();
+            RollbarHelper.ConfigureRollbarSingleton();
 
             RollbarLocator.RollbarInstance
                 .Info("Xamarin.Forms sample: Hello world! Xamarin is here...");
 
+            //START: Let's subscribe to all known unhandled exception events application-wide...
+            RollbarHelper.RegisterForGlobalExceptionHandling();
             this.UnhandledException += App_UnhandledException;
+            //END.
 
             this.InitializeComponent();
             this.Suspending += OnSuspending;
@@ -45,7 +48,7 @@ namespace Sample.Xamarin.Forms.UWP
 
         private void App_UnhandledException(object sender, Windows.UI.Xaml.UnhandledExceptionEventArgs e)
         {
-            RollbarLocator.RollbarInstance.AsBlockingLogger(TimeSpan.FromSeconds(5)).Critical(e.Exception);
+            RollbarLocator.RollbarInstance.AsBlockingLogger(RollbarHelper.RollbarTimeout).Critical(e.Exception);
         }
 
         /// <summary>
@@ -114,86 +117,6 @@ namespace Sample.Xamarin.Forms.UWP
             var deferral = e.SuspendingOperation.GetDeferral();
             //TODO: Save application state and stop any background activity
             deferral.Complete();
-        }
-
-        /// <summary>
-        /// Configures the Rollbar singleton-like notifier.
-        /// </summary>
-        private static void ConfigureRollbarSingleton()
-        {
-            const string rollbarAccessToken = "17965fa5041749b6bf7095a190001ded";
-            const string rollbarEnvironment = "RollbarNetSamples";
-
-            var config = new RollbarConfig(rollbarAccessToken) // minimally required Rollbar configuration
-            {
-                Environment = rollbarEnvironment,
-                ScrubFields = new string[]
-                {
-                    "access_token", // normally, you do not want scrub this specific field (it is operationally critical), but it just proves safety net built into the notifier... 
-                    "username",
-                }
-            };
-            RollbarLocator.RollbarInstance
-                // minimally required Rollbar configuration:
-                .Configure(config)
-                // optional step if you would like to monitor this Rollbar instance's internal events within your application:
-                .InternalEvent += OnRollbarInternalEvent
-                ;
-
-            // optional step if you would like to monitor all Rollbar instances' internal events within your application:
-            RollbarQueueController.Instance.InternalEvent += OnRollbarInternalEvent;
-
-            // Optional info about reporting Rollbar user:
-            SetRollbarReportingUser("007", "jbond@mi6.uk", "JBOND");
-        }
-
-        /// <summary>
-        /// Sets the rollbar reporting user.
-        /// </summary>
-        /// <param name="id">The identifier.</param>
-        /// <param name="email">The email.</param>
-        /// <param name="userName">Name of the user.</param>
-        private static void SetRollbarReportingUser(string id, string email, string userName)
-        {
-            rdto.Person person = new rdto.Person(id);
-            person.Email = email;
-            person.UserName = userName;
-            RollbarLocator.RollbarInstance.Config.Person = person;
-        }
-
-        /// <summary>
-        /// Called when rollbar internal event is detected.
-        /// </summary>
-        /// <param name="sender">The sender.</param>
-        /// <param name="e">The <see cref="RollbarEventArgs"/> instance containing the event data.</param>
-        private static void OnRollbarInternalEvent(object sender, RollbarEventArgs e)
-        {
-            Console.WriteLine(e.TraceAsString());
-
-            RollbarApiErrorEventArgs apiErrorEvent = e as RollbarApiErrorEventArgs;
-            if (apiErrorEvent != null)
-            {
-                //TODO: handle/report Rollbar API communication error event...
-                return;
-            }
-            CommunicationEventArgs commEvent = e as CommunicationEventArgs;
-            if (commEvent != null)
-            {
-                //TODO: handle/report Rollbar API communication event...
-                return;
-            }
-            CommunicationErrorEventArgs commErrorEvent = e as CommunicationErrorEventArgs;
-            if (commErrorEvent != null)
-            {
-                //TODO: handle/report basic communication error while attempting to reach Rollbar API service... 
-                return;
-            }
-            InternalErrorEventArgs internalErrorEvent = e as InternalErrorEventArgs;
-            if (internalErrorEvent != null)
-            {
-                //TODO: handle/report basic internal error while using the Rollbar Notifier... 
-                return;
-            }
         }
 
     }
